Nomes utlizadores nao aparecem nas tarefas

+++ GPS_REST/models/participa.js
@@ -21,7 +21,7 @@ module.exports = function(sequelize, DataTypes) {
       }
     },
     acepted: {
-      type: DataTypes.INTEGER(1),
+      type: DataTypes.INTEGER(4),
       allowNull: false
     }
   }, {
